Created Depraction Warning
At time of writing, this repo is depracted - It was the first of our repos and therefore doesn´t contain any of our new, refreshed repos, including @RaiseJS/cli, which is still called VJQ and isn´t nearly as extendable as the new core. Therefore, we do NOT recommend using this tool until a major update is released.